fix(flutter): keep create task advanced options collapsed
